selector value appears dirrectly on table
